specify explicit size for icon svgs
Sometimes you can see a 'flash' of a very large svg image while the page is loading. This is because the svg has loaded before the css specifying its size has. To fix, specify a reasonable size explicitly in the svg (so it will load at that size by default and then been overridden by whatever size the css tells it to be.